# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U-ED News API Documentation",
        "description": "A simple API made with Laravel 8",
        "contact": {
            "email": "work.jerson.baltazar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get all Tags",
                "description": "Returns list of tag data",
                "operationId": "indexTag",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Filter by the slug of tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by the name of tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status of article 1 = Active, 0 = Inactive",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "15"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Store new tag",
                "description": "Returns tag data",
                "operationId": "storeTags",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTagsRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Show certain tag",
                "description": "Returns an tag data",
                "operationId": "showTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tags id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update existing tag",
                "description": "Returns updated tag data",
                "operationId": "updateTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tags id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTagsRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete existing Tags",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreTagsRequestModel": {
                "title": "Store Tag request",
                "description": "Store Tag request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of the new article",
                        "type": "string",
                        "example": "Sample Tag"
                    }
                },
                "type": "object"
            },
...
            "UpdateTagsRequestModel": {
                "title": "Update Tag request",
                "description": "Update Tag request body data",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreTagsRequestModel"
                    },
                    {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "status": {
                                "title": "status",
                                "description": "status of the tag  1 = Active, 0 = Inactive",
                                "type": "string",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}